Changed x to a UINT16

+++ internal/internal.h
@@ -25,7 +25,7 @@ typedef struct {
 void changeLevel(UINT8 level);
 
 // scrolling
-extern bool scroll(UINT8 player_x, UINT8 x_mod, UINT8 y_mod, UINT8 *pixels_scrolled, UINT8 *tiles_scrolled);
+extern bool scroll(UINT8 player_x, UINT16 x_mod, UINT8 y_mod, UINT8 *pixels_scrolled, UINT8 *tiles_scrolled);
 
 // life
 extern void init_hp();
