+++ movement/collision.c
@@ -5,7 +5,7 @@
 
 // Check if two rectangles from x1,y1, and extending out w1, h2, 
 // overlap with another, x2,y2, and extending out w2, h2
-bool collision_check(UINT8 x1, UINT8 y1, UINT8 w1, UINT8 h1, UINT8 x2, UINT8 y2, UINT8 w2, UINT8 h2) {
+bool collision_check(UINT16 x1, UINT8 y1, UINT8 w1, UINT8 h1, UINT16 x2, UINT8 y2, UINT8 w2, UINT8 h2) {
 	if ((x1 < (x2+w2)) && ((x1+w1) > x2) && (y1 < (h2+y2)) && ((y1+h1) > y2)) {
 		return true;
 	} else {
@@ -13,7 +13,7 @@ bool collision_check(UINT8 x1, UINT8 y1, UINT8 w1, UINT8 h1, UINT8 x2, UINT8 y2,
 	}
 }
 
-bool player_collision_with_screen(UINT8 x, UINT8 y, UINT8 player_size, UINT8 tile_size){
+bool player_collision_with_screen(UINT16 x, UINT8 y, UINT8 player_size, UINT8 tile_size){
 	if(
 		// left wall
 		collision_check(x, y, player_size, player_size, 8, 16, 0, 144)
@@ -31,7 +31,7 @@ bool player_collision_with_screen(UINT8 x, UINT8 y, UINT8 player_size, UINT8 til
 	return false;
 }
 
-bool player_collision_with_obstacles(UINT8 x, UINT8 y, UINT8 player_size, Obstacle *obstacles, UINT8 obstacle_count){
+bool player_collision_with_obstacles(UINT16 x, UINT8 y, UINT8 player_size, Obstacle *obstacles, UINT8 obstacle_count){
 	for(UINT8 i = 0; i < obstacle_count; i++){
 		if(
 			collision_check(x, y, player_size, player_size, obstacles[i].x, obstacles[i].y, obstacles[i].width, obstacles[i].height)
@@ -43,7 +43,7 @@ bool player_collision_with_obstacles(UINT8 x, UINT8 y, UINT8 player_size, Obstac
 	return false;
 }
 
-bool player_collision_with_enemies(UINT8 x, UINT8 y, UINT8 player_size, Enemy *enemies, UINT8 enemy_count){
+bool player_collision_with_enemies(UINT16 x, UINT8 y, UINT8 player_size, Enemy *enemies, UINT8 enemy_count){
 	for(UINT8 i = 0; i < enemy_count; i++){
 		if (enemies[i].active == 0) { continue; }
